Logged-in Performance Profiler: Add loading state
Logged-in Performance Profiler: Add loading state
