+++ client/hosting/performance/components/PerformanceReport.tsx
@@ -1,29 +1,25 @@
-import { useUrlBasicMetricsQuery } from 'calypso/data/site-profiler/use-url-basic-metrics-query';
-import { useUrlPerformanceInsightsQuery } from 'calypso/data/site-profiler/use-url-performance-insights';
+import { PerformanceReport as PerformanceReportObject } from 'calypso/data/site-profiler/types';
 import { PerformanceProfilerDashboardContent } from 'calypso/performance-profiler/components/dashboard-content';
-import { Tab } from './device-tab-control';
+import { PerformanceReportLoading } from './PerformanceReportLoading';
 
 interface PerformanceReportProps {
-	wpcom_performance_url?: { url: string; hash: string };
-	activeTab: Tab;
+	performanceReport?: PerformanceReportObject;
+	url: string;
+	hash: string;
+	isLoading: boolean;
+	pageTitle: string;
 }
 
 export const PerformanceReport = ( {
-	wpcom_performance_url,
-	activeTab,
+	isLoading,
+	performanceReport,
+	url,
+	hash,
+	pageTitle,
 }: PerformanceReportProps ) => {
-	const { url = '', hash = '' } = wpcom_performance_url || {};
-
-	const { data: basicMetrics } = useUrlBasicMetricsQuery( url, hash, true );
-	const { final_url: finalUrl } = basicMetrics || {};
-	const { data: performanceInsights } = useUrlPerformanceInsightsQuery( url, hash );
-
-	const mobileReport =
-		typeof performanceInsights?.mobile === 'string' ? undefined : performanceInsights?.mobile;
-	const desktopReport =
-		typeof performanceInsights?.desktop === 'string' ? undefined : performanceInsights?.desktop;
-
-	const performanceReport = activeTab === 'mobile' ? mobileReport : desktopReport;
+	if ( isLoading ) {
+		return <PerformanceReportLoading isSavedReport={ !! hash } pageTitle={ pageTitle } />;
+	}
 
 	if ( ! performanceReport ) {
 		return null;
@@ -32,7 +28,7 @@ export const PerformanceReport = ( {
 	return (
 		<PerformanceProfilerDashboardContent
 			performanceReport={ performanceReport }
-			url={ finalUrl ?? url }
+			url={ url }
 			hash={ hash }
 			displayThumbnail={ false }
 			displayNewsletterBanner={ false }
